refactor: Chip 컴포넌트 shared의 하위 경로로 이동
